requirements.txt
Django
djangorestframework
psycopg2-binary
celery
redis
gunicorn
django-cors-headers
django-environ
drf-yasg
pytest
pytest-django
coverage
django-filter
djangorestframework-simplejwt
